Split out emitting _start into dedicated pass
