+++ mlir/include/qcc/Conversion/ToIntrinsics/ToIntrinsics.td
@@ -36,10 +36,6 @@ def ConvertQIRToIntrinsics : Pass<"convert-qir-to-intrinsics", "mlir::ModuleOp">
 
       Leftover QIR function declarations are removed from the module.
 
-      Additionally, if a function carries the `entry_point` passthrough
-      attribute, a `_start` function is emitted that sets up the stack
-      and calls it, superseding it as the hardware's actual entry point.
-
       Assumptions:
       - The module has been fully lowered by the ToQIR pipeline
         (`prep-to-qir`, `convert-qc-to-qir`, `finalize-to-qir`) so all gate
@@ -47,6 +43,30 @@ def ConvertQIRToIntrinsics : Pass<"convert-qir-to-intrinsics", "mlir::ModuleOp">
       - Qubit pointers were produced by `llvm.inttoptr(llvm.mlir.constant(N : i64))`,
         i.e. the static qubit management scheme used by `convert-qc-to-qir`.
    }];
+
+  let dependentDialects = [
+    "mlir::LLVM::LLVMDialect"
+  ];
+}
+
+def EmitHiSEPQStart : Pass<"emit-hisepq-start", "mlir::ModuleOp"> {
+  let summary = "Emits the HiSEP-Q `_start` entry point.";
+
+  let description = [{
+      Appends a `_start` function that supersedes the module's entry point as
+      the entry point of the hardware. It sets the stack pointer to
+      `__stack_top`, jumps to the original entry point and loops forever when
+      that returns.
+
+      Both symbol names are contractual with the HiSEP-Q linker script
+      (`hisepq.ld`), which declares `ENTRY(_start)`, pulls `.text._start` to
+      the boot address and provides `__stack_top`.
+
+      Assumes that *exactly* one function is tagged entry point.
+
+      TODO: Drop this pass once HiSEP-Q ships a crt0. Then qcc can emit a plain
+      `main` and leave the startup code to the runtime.
+  }];
 
   let dependentDialects = [
     "mlir::LLVM::LLVMDialect"
