+++ mlir/lib/Conversion/ToIntrinsics/ConvertQIRToIntrinsics.cpp
@@ -10,17 +10,14 @@
 #include "qcc/Conversion/ToQIR/Constants.h"
 
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
-#include "mlir/Dialect/LLVMIR/LLVMTypes.h"
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/IR/SymbolTable.h"
 #include "mlir/IR/Value.h"
-#include "mlir/IR/ValueRange.h"
 #include "mlir/Pass/Pass.h" // FIXME: why does clangd complain about unused include?
 #include "mlir/Support/LogicalResult.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 
-#include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/ADT/StringMap.h"
 #include "llvm/ADT/StringRef.h"
@@ -258,19 +255,6 @@ namespace qcc {
 #define GEN_PASS_DEF_CONVERTQIRTOINTRINSICS
 #include "qcc/Conversion/ToIntrinsics/ToIntrinsics.h.inc"
 
-/// Whether `funcOp` carries the `entry_point` passthrough attribute.
-static bool isEntryPointFunc(LLVM::LLVMFuncOp funcOp) { // FIXME: find a better place
-  auto passthrough = funcOp->getAttrOfType<ArrayAttr>("passthrough");
-  if (!passthrough) {
-    return false;
-  }
-
-  return llvm::any_of(passthrough, [](Attribute attr) {
-    auto strAttr = dyn_cast<StringAttr>(attr);
-    return strAttr && strAttr.getValue() == "entry_point";
-  });
-}
-
 namespace {
 
 struct ConvertQIRToIntrinsics final : impl::ConvertQIRToIntrinsicsBase<ConvertQIRToIntrinsics> {
@@ -291,42 +275,9 @@ protected:
     }
 
     removeUnusedQIRSymbols();
-
-    FailureOr<LLVM::LLVMFuncOp> entryPoint = getEntryPoint(moduleOp);
-    if (failed(entryPoint)) {
-      return signalPassFailure();
-    }
-
-    if (*entryPoint) {
-      emitStartFunc(moduleOp, *entryPoint);
-    }
   }
 
 private:
-  /// Returns the entry point of the module. Fails if not exactly one function
-  /// is tagged, as the hardware boots at a single address.
-  static FailureOr<LLVM::LLVMFuncOp> getEntryPoint(ModuleOp moduleOp) {
-    LLVM::LLVMFuncOp entryPoint;
-
-    for (auto funcOp : moduleOp.getOps<LLVM::LLVMFuncOp>()) {
-      if (!isEntryPointFunc(funcOp)) {
-        continue;
-      }
-      if (entryPoint) {
-        return funcOp.emitError("expected at most one function tagged as the entry point, but found '")
-               << entryPoint.getName() << "' and '" << funcOp.getName()
-               << "'"; // FIXME: if funcOp emits the error do we really need to print the name?
-      }
-      entryPoint = funcOp;
-    }
-
-    if (!entryPoint) {
-      return moduleOp->emitError("did not find any entry point");
-    }
-
-    return entryPoint;
-  }
-
   /// Removes QIR function declarations and globals that have no remaining uses.
   ///
   /// Besides the `__quantum__*` declarations this also drops the dummy output
@@ -351,35 +302,6 @@ private:
     for (auto* op : toErase) {
       op->erase();
     }
-  }
-
-  /// Emits `_start`, which supersedes `entryPoint` as the entry point of the hardware.
-  static void emitStartFunc(ModuleOp moduleOp,
-                            LLVM::LLVMFuncOp entryPoint) { // FIXME: in-depth understanding of this, and testing.
-    OpBuilder builder(moduleOp.getContext());
-    builder.setInsertionPointToEnd(moduleOp.getBody());
-    Location loc = entryPoint.getLoc();
-
-    auto stackTopType = LLVM::LLVMArrayType::get(builder.getI8Type(), 0);
-    auto stackTop = LLVM::GlobalOp::create(builder, loc, stackTopType, /*isConstant=*/true, LLVM::Linkage::External,
-                                           "__stack_top", /*value=*/Attribute());
-
-    auto startFuncType = LLVM::LLVMFunctionType::get(LLVM::LLVMVoidType::get(builder.getContext()), {});
-    auto startFunc = LLVM::LLVMFuncOp::create(builder, loc, "_start", startFuncType);
-    builder.setInsertionPointToStart(startFunc.addEntryBlock(builder));
-
-    Value stackTopAddr = LLVM::AddressOfOp::create(builder, loc, stackTop);
-    Value entryAddr = LLVM::AddressOfOp::create(builder, loc, entryPoint);
-
-    auto asmDialect = LLVM::AsmDialectAttr::get(builder.getContext(), LLVM::AsmDialect::AD_ATT);
-    LLVM::InlineAsmOp::create(builder, loc, /*resultTypes=*/TypeRange(),
-                              /*operands=*/ValueRange{stackTopAddr, entryAddr},
-                              /*asm_string=*/"mv sp, $0\njalr ra, 0($1)\n1:\nj 1b",
-                              /*constraints=*/"r,r", /*has_side_effects=*/true,
-                              /*is_align_stack=*/false, LLVM::TailCallKind::None,
-                              /*asm_dialect=*/asmDialect, /*operand_attrs=*/ArrayAttr());
-
-    LLVM::UnreachableOp::create(builder, loc);
   }
 };
