+++ mlir/test/tools/qcc-opt/ToIntrinsics/convert-qir-to-intrinsics-error.mlir
@@ -39,17 +39,3 @@ llvm.func @too_few_qubit_operands() {
   llvm.call @__quantum__qis__cx__body(%q) : (!llvm.ptr) -> ()
   llvm.return
 }
-
-// -----
-
-// The hardware boots at a single address, so at most one function may be tagged
-// as the entry point.
-
-llvm.func @first_entry() attributes { passthrough = ["entry_point"] } {
-  llvm.return
-}
-
-// expected-error @+1 {{expected at most one function tagged as the entry point, but found 'first_entry' and 'second_entry'}}
-llvm.func @second_entry() attributes { passthrough = ["entry_point"] } {
-  llvm.return
-}
